- move code and templates and create console script definitions
