+++ CHANGES.rst
@@ -1,6 +1,5 @@
-=========
- Changes
-=========
+Change log
+==========
 
 1.0 (unreleased)
-================
+----------------
