--- config/requirements.txt
@@ -1,8 +1,0 @@
-check-python-versions==0.22.0
-Jinja2==3.1.4
-packaging==24.1
-pyupgrade==3.16.0
-requests==2.32.3
-tomlkit==0.13.2
-tox==4.18.1
-zest.releaser==9.2.0
