+++ src/zope/meta/default/MANIFEST.in.j2
