+++ src/zope/meta/default/setup.cfg.j2
