+++ src/zope/meta/multi_call.py
@@ -14,10 +14,10 @@
 import argparse
 import sys
 
-from shared.call import call
-from shared.packages import list_packages
-from shared.path import change_dir
-from shared.path import path_factory
+from .shared.call import call
+from .shared.packages import list_packages
+from .shared.path import change_dir
+from .shared.path import path_factory
 
 
 parser = argparse.ArgumentParser(
