+++ src/zope/meta/re_enable_actions.py
@@ -14,9 +14,9 @@
 import argparse
 import pathlib
 
-from shared.call import call
-from shared.packages import ALL_REPOS
-from shared.packages import ORG
+from .shared.call import call
+from .shared.packages import ALL_REPOS
+from .shared.packages import ORG
 
 
 base_url = f'https://github.com/{ORG}'
