+++ src/zope/meta/shared/git.py
