+++ src/zope/meta/shared/path.py
